Added implementation of instructor-exception

+++ src/main/java/se/yrgo/dataaccess/InstructorDaoJpaImpl.java
@@ -5,6 +5,7 @@ import se.yrgo.domain.GymClass;
 import se.yrgo.domain.Instructor;
 
 import javax.persistence.EntityManager;
+import javax.persistence.NoResultException;
 import javax.persistence.PersistenceContext;
 import java.util.List;
 
@@ -19,21 +20,32 @@ public class InstructorDaoJpaImpl implements InstructorDao {
     }
 
     @Override
-    public void modifyInstructor(Instructor changedInstructor) {
+    public void modifyInstructor(Instructor changedInstructor) throws InstructorNotFoundException {
+        Instructor existing = em.find(Instructor.class, changedInstructor.getInstructorId());
+        if (existing == null) {
+            throw new InstructorNotFoundException("Instructor not found");
+        }
         em.merge(changedInstructor);
     }
 
     @Override
-    public void removeInstructor(Instructor deletedInstructor) {
+    public void removeInstructor(Instructor deletedInstructor) throws InstructorNotFoundException {
         Instructor instructor = em.find(Instructor.class, deletedInstructor.getInstructorId());
+        if (instructor == null) {
+            throw new InstructorNotFoundException("Instructor not found");
+        }
         em.remove(instructor);
     }
 
     @Override
-    public Instructor getInstructorById(String id) throws Exception {
-        return em.createQuery("select i from Instructor as i where i.instructorId =:id", Instructor.class)
-                .setParameter("id", id)
-                .getSingleResult();
+    public Instructor getInstructorById(String id) throws InstructorNotFoundException {
+        try {
+            return em.createQuery("select i from Instructor as i where i.instructorId =:id", Instructor.class)
+                    .setParameter("id", id)
+                    .getSingleResult();
+        } catch (NoResultException e) {
+            throw new InstructorNotFoundException("Instructor with id: " + id + " not found");
+        }
     }
 
     @Override
@@ -49,7 +61,7 @@ public class InstructorDaoJpaImpl implements InstructorDao {
     }
 
     @Override
-    public List<GymClass> getGymClasses(String id) {
+    public List<GymClass> getGymClasses(String id) throws InstructorNotFoundException {
         return em.createQuery("select g from GymClass as g where g.instructor.instructorId =:id", GymClass.class)
                 .setParameter("id", id)
                 .getResultList();
@@ -57,7 +69,7 @@ public class InstructorDaoJpaImpl implements InstructorDao {
 
     @Override
     public int getNumberOfClassesForInstructor(Instructor instructor) {
-        return  (int) em.createQuery("select count(g) FROM GymClass as g WHERE g.instructor = :instructor")
+        return (int) em.createQuery("select count(g) FROM GymClass as g WHERE g.instructor = :instructor")
                 .setParameter("instructor", instructor)
                 .getSingleResult();
     }
